Updated README and other documentation
