+++ raid-agency-app-static/scripts/fetch-raids.js
@@ -59,7 +59,8 @@ import http from 'http';
 import { config as dotenvConfig } from 'dotenv';
 import { fileURLToPath } from 'url';
 import { dirname } from 'path';
-import { fetchAllHandles } from './fetch-handles.js';
+import { fetchCitation } from './fetch-citation.js'
+import fetchAllHandles, { extractHandles } from './fetch-handles.js';
 
 // Get __dirname equivalent in ES modules
 const __filename = fileURLToPath(import.meta.url);
@@ -114,7 +115,7 @@ function validateConfig() {
 }
 
 // Enhanced HTTP request with retry logic
-async function makeRequestWithRetry(url, options = {}, retries = config.maxRetries) {
+export async function makeRequestWithRetry(url, options = {}, retries = config.maxRetries) {
   for (let attempt = 1; attempt <= retries; attempt++) {
     try {
       const response = await makeRequest(url, {
@@ -229,85 +230,12 @@ async function fetchRaidData(bearerToken) {
   }
 }
 
-function cleanDetailedCitation(citation) {
-  let cleaned = citation;
-
-  // Remove DOI URLs (both http and https, with or without dx.)
-  // This regex matches URLs that start with http(s)://doi.org or dx.doi.org
-  // and continues until it hits whitespace or end of string
-  cleaned = cleaned.replace(/https?:\/\/(?:dx\.)?doi\.org\/[^\s]+/gi, '');
-
-  // Remove standalone DOI patterns (e.g., "DOI: 10.1234/..." or just "10.1234/...")
-  // that might appear anywhere in the text
-  cleaned = cleaned.replace(/\bDOI\s*:?\s*10\.\d+\/[^\s,.;]+/gi, '');
-  
-  // Remove bare DOI patterns that start with 10. and contain a forward slash
-  cleaned = cleaned.replace(/\b10\.\d+\/[^\s,.;]+/g, '');
-
-  // Clean up any trailing punctuation that might be left after DOI removal
-  cleaned = cleaned.replace(/\s*[,;]\s*\./, '.');
-  
-  // Clean up extra spaces
-  cleaned = cleaned.replace(/\s{2,}/g, ' ');
-  
-  // Clean up any double periods
-  cleaned = cleaned.replace(/\.\s*\.\s*/g, '. ');
-  
-  // Remove any trailing whitespace before periods
-  cleaned = cleaned.replace(/\s+\./g, '.');
-
-  // Ensure proper ending
-  cleaned = cleaned.trim();
-  if (cleaned && !cleaned.endsWith('.') && !cleaned.endsWith('!') && !cleaned.endsWith('?')) {
-    cleaned += '.';
-  }
-
-  return cleaned;
-}
-
-// Fetch citation for a DOI with caching
-async function fetchCitation(doi) {
-  // Check cache first
-  if (config.enableCaching && citationCache.has(doi) && (Date.now() - citationCache.get(doi).timestamp < config.cachingTime)) {
-    stats.cachedCitations++;
-    return citationCache.get(doi).citation;
-  }
-  
-  const url = `${doi}`;
-  
-  try {
-    const response = await makeRequestWithRetry(url, {
-      method: 'GET',
-      headers: {
-        'Accept': 'text/bibliography; style=apa, text/x-bibliography; style=apa, text/plain'
-      }
-    }, 2); // Fewer retries for DOI lookups
-    
-    // Check if we got a valid citation
-    const citation = cleanDetailedCitation(response.data);
-
-    if (citation && !citation.includes('DOI Not Found')) {
-      if (config.enableCaching) {
-        citationCache.set(doi, {
-          citation: citation,
-          timestamp: Date.now()
-        });
-      }
-      stats.successfulCitations++;
-      return citation;
-    }
-    return "DOI exists but metadata could not be retrieved - contact info@doi.org for help with this";
-  } catch (error) {
-    stats.failedCitations++;
-    return "DOI exists but metadata could not be retrieved - contact info@doi.org for help with this";
-  }
-}
-
 // Process a batch of DOIs concurrently
 async function processDOIBatch(dois) {
   const results = await Promise.all(
     dois.map(async ({ doi, raidIndex, objectIndex }) => {
-      const citation = await fetchCitation(doi);
+      const params = {doi, makeRequestWithRetry, stats, citationCache, config}
+      const citation = await fetchCitation(params);
       return { citation, raidIndex, objectIndex };
     })
   );
@@ -384,17 +312,6 @@ async function addCitationsToRaidData(raidData) {
   return raidData;
 }
 
-// Save cache to file (if enabled)
-async function saveCache() {
-  if (config.enableCaching && citationCache.size > 0) {
-    const cacheFile = path.join(config.dataDir, '.citation-cache.json');
-    const cacheData = Object.fromEntries(citationCache);
-    await fs.writeFile(cacheFile, JSON.stringify(cacheData, null, 2));
-    console.log(`Citation cache saved (${citationCache.size} entries)`);
-  }
-}
-
-// Load cache from file (if enabled and exists)
 async function loadCache() {
   if (config.enableCaching) {
     const cacheFile = path.join(config.dataDir, '.citation-cache.json');
@@ -411,6 +328,16 @@ async function loadCache() {
   }
 }
 
+// Load cache from file (if enabled and exists)
+// Save cache to file (if enabled)
+async function saveCache() {
+  if (config.enableCaching && citationCache.size > 0) {
+    const cacheFile = path.join(config.dataDir, '.citation-cache.json');
+    const cacheData = Object.fromEntries(citationCache);
+    await fs.writeFile(cacheFile, JSON.stringify(cacheData, null, 2));
+    console.log(`Citation cache saved (${citationCache.size} entries)`);
+  }
+}
 // Main execution
 async function main() {
   try {
@@ -439,7 +366,8 @@ async function main() {
     console.log(`Data successfully saved to ${outputFile}`);
     
     // Step 5: Extract and save handles
-    const handles = await fetchAllHandles();
+    await fetchAllHandles();
+    const handles = await extractHandles(raidData);
     const handlesFile = path.join(config.dataDir, 'handles.json');
     await fs.writeFile(handlesFile, JSON.stringify(handles, null, 2));
     console.log(`Unique handles saved to ${handlesFile}`);
@@ -478,5 +406,4 @@ export {
   fetchRaidData,
   addCitationsToRaidData,
   fetchCitation,
-  config
 };
