+++ raid-agency-app-static/src/components/raid-components/related-objects.astro
@@ -29,7 +29,7 @@ const { relatedObjects } = Astro.props as Props;
           return (
             <div class="ml-5 space-y-2 py-4">
               <div class="grid grid-cols-2 gap-5 text-sm">
-                <InfoField label="Citation" value={el.citation?.text || ""} />
+                {/* <InfoField label="Citation" value={el.citation?.text || ""} /> */}
                 <InfoField label="ID" value={el.id || ""} />
                 <InfoField label="Type" value={relatedObjectType || ""} />
               </div>
